requirements.txt
ultralytics==8.1.0
torch
torchvision
torchaudio

transformers==4.39.0
Pillow
opencv-python
numpy
regex

# Optional but recommended
tqdm
